Disallow relative paths after authority parts.
This allows us to adhere to the URI standard by disallowing a second slash to
qualify a full path. Instead, engines that like paths always get a full path,
and engines that don't use paths just get the database name (unless it looks
like a Windows path, in which case all bets are off).

Might end up reverting this change, though, because some engines might take
either a file path *or* a database name, and there isn't a way to tell the
difference. I assert that this is a weakness of the URI standard, however, and
think allowing the double-slash is the proper way to go for a situation like
this.