Revert "Disallow relative paths after authority parts."
Turns out that RFC 3986 allows a double slash at the start of the path part.
This greatly simplifies things for us, as we can easily declare a relative or
absolute path, or a database name, without regard to the slash that separates
the path from the authority. In other words, the first slash separates the
authority from the path, but any other slashes after that, including
immediately afterward, are part of the path.

This reverts commit 0fdfa078da8be51def01554deed0be6524658de1.

Conflicts:
	README.md