Turn URI::db into a wrapper around a URI::_db object.
So that those database URIs can be used on their own, with the db: prefix.
Requires that URI::db inherit from *something*, or else URI.pm can't see it,
so using URI::WithBase for now.

In order for the known database URIs to be used without the db: prefix,
move them out of the db/ subdirectory, so that URI.pm can find them.