Fix dbname for URI with // but no authority.
In "db:sqlite:///foo.db", the `dbname` should be "foo.db", not "/foo.db". So
try harder to detect that situation. Instead of looking for no slash afte the
two slashes, look instead for a lack of *two* slashes. Yes this is weird. I
find myself again wishing that URI objects were deparsed, instead of stored as
a string to be parsed in every method. Resolves #8.