Add test and release workflows
Remove Travis CI config, which wasn't working anymore. The test workflow
tests Perls 5.8 - 5.36 on Linux, macOS, and Windows. The release
workflow uses a custom Build.PL targets to build a list of changes and
to print the name of the tarball for releasing on GitHub. Requires
GitHub secrets `CPAN_PASSWORD` and `CPAN_USERNAME` to release on CPAN.